requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
# colpali_engine==0.3.14.dev2
comm==0.2.3
configue==6.0.0
datasets==2.21.0
debugpy==1.8.17
decorator==5.2.1
dill==0.3.8
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.19.1
flash-attn==2.7.3
frozenlist==1.8.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
GPUtil==1.4.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==6.31.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.3.2
parso==0.8.5
pdf2image==1.17.0
peft==0.14.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.4.0
polars==1.35.2
polars-runtime-32==1.35.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pynvml==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytrec_eval-terrier==0.5.7
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
sentencepiece==0.2.0
sentry-sdk==2.44.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.6.0
torchvision==0.21.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.2.0
typer==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
# vidore_benchmark==5.0.0
wandb==0.23.0
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0